Remove protocol.go
